+++ cartridges/app_stripe_sfra/cartridge/client/default/js/stripe.checkout.js
@@ -563,7 +563,21 @@ function handleStripePaymentElementSubmitOrder() {
     });
 }
 
-function initStripePaymentElement() {
+function getBillingDetails(ownerEmail) {
+    return {
+        billingDetails: {
+            email: ownerEmail,
+            name: (document.querySelector('#dwfrm_billing input[name$="_firstName"]') && document.querySelector('#dwfrm_billing input[name$="_lastName"]')) ? document.querySelector('#dwfrm_billing input[name$="_firstName"]').value + ' ' + document.querySelector('#dwfrm_billing input[name$="_lastName"]').value : '',
+            phone: document.querySelector('#dwfrm_billing input[name$="_phone"]') ? document.querySelector('#dwfrm_billing input[name$="_phone"]').value : '',
+            address: {
+                postal_code: document.querySelector('#dwfrm_billing input[name$="_postalCode"]') ? document.querySelector('#dwfrm_billing input[name$="_postalCode"]').value : '',
+                country: document.querySelector('#dwfrm_billing select[name$="_country"]') ? document.querySelector('#dwfrm_billing select[name$="_country"]').value : ''
+            }
+        }
+    };
+}
+
+function initStripePaymentElement(customerEmail) {
     var ownerEmail = '';
     if ($('.customer-summary-email').length && $('.customer-summary-email').text()) {
         ownerEmail = $('.customer-summary-email').text();
@@ -581,85 +595,32 @@ function initStripePaymentElement() {
 
         if (window.stripePaymentElements) {
             window.paymentElementInstance = window.stripePaymentElements.create('payment', {
-                defaultValues: {
-                    billingDetails: {
-                        email: ownerEmail,
-                        name: (document.querySelector('#dwfrm_billing input[name$="_firstName"]') && document.querySelector('#dwfrm_billing input[name$="_lastName"]')) ? document.querySelector('#dwfrm_billing input[name$="_firstName"]').value + ' ' + document.querySelector('#dwfrm_billing input[name$="_lastName"]').value : '',
-                        phone: document.querySelector('#dwfrm_billing input[name$="_phone"]') ? document.querySelector('#dwfrm_billing input[name$="_phone"]').value : '',
-                        address: {
-                            postal_code: document.querySelector('#dwfrm_billing input[name$="_postalCode"]') ? document.querySelector('#dwfrm_billing input[name$="_postalCode"]').value : '',
-                            country: document.querySelector('#dwfrm_billing select[name$="_country"]') ? document.querySelector('#dwfrm_billing select[name$="_country"]').value : ''
-                        }
-                    }
-                }
+                defaultValues: getBillingDetails(ownerEmail)
             });
             window.paymentElementInstance.mount('#payment-element');
         }
     } else {
-        $.ajax({
-            url: document.getElementById('getCustomerEmailURL').value,
-            method: 'GET',
-            dataType: 'json',
-            async: false
-        }).done(function (json) {
-            ownerEmail = json.email;
-            if (ownerEmail && ownerEmail !== 'null') {
-                if (window.paymentElementInstance) {
-                    window.paymentElementInstance.destroy();
-                }
+        if (window.paymentElementInstance) {
+            window.paymentElementInstance.destroy();
+        }
 
-                window.paymentElementInstance = window.stripePaymentElements.create('payment', {
-                    defaultValues: {
-                        billingDetails: {
-                            email: ownerEmail,
-                            name: (document.querySelector('#dwfrm_billing input[name$="_firstName"]') && document.querySelector('#dwfrm_billing input[name$="_lastName"]')) ? document.querySelector('#dwfrm_billing input[name$="_firstName"]').value + ' ' + document.querySelector('#dwfrm_billing input[name$="_lastName"]').value : '',
-                            phone: document.querySelector('#dwfrm_billing input[name$="_phone"]') ? document.querySelector('#dwfrm_billing input[name$="_phone"]').value : '',
-                            address: {
-                                postal_code: document.querySelector('#dwfrm_billing input[name$="_postalCode"]') ? document.querySelector('#dwfrm_billing input[name$="_postalCode"]').value : '',
-                                country: document.querySelector('#dwfrm_billing select[name$="_country"]') ? document.querySelector('#dwfrm_billing select[name$="_country"]').value : ''
-                            }
-                        }
-                    }
-                });
-                window.paymentElementInstance.mount('#payment-element');
-            }
+        window.paymentElementInstance = window.stripePaymentElements.create('payment', {
+            defaultValues: getBillingDetails(customerEmail)
         });
+        window.paymentElementInstance.mount('#payment-element');
     }
 }
 
 function initNewStripePaymentIntent() {
-    var stripeOrderAmountInput = document.getElementById('stripe_order_amount');
-    if (!stripeOrderAmountInput || !stripeOrderAmountInput.value) {
-        return;
-    }
+    $.ajax({
+        url: document.getElementById('getPaymentElementOptions').value,
+        method: 'GET',
+        dataType: 'json',
+    }).done(function (response) {
+        window.stripePaymentElements = stripe.elements(response.paymentElementOptions);
 
-    var stripeOrderCurrencyInput = document.getElementById('stripe_order_currency');
-    if (!stripeOrderCurrencyInput || !stripeOrderCurrencyInput.value) {
-        return;
-    }
-
-    const appearance = {
-        theme: 'stripe'
-    };
-
-    const stripePaymentElementStyleObject = JSON.parse(document.getElementById('stripePaymentElementStyle').value);
-    appearance.variables = stripePaymentElementStyleObject.variables;
-
-    const options = {
-        mode: 'payment',
-        amount: parseInt(stripeOrderAmountInput.value, 10),
-        currency: stripeOrderCurrencyInput.value,
-        appearance: appearance,
-        capture_method: document.getElementById('stripeCaptureMethod').value,
-    };
-
-    if (document.getElementById('isStripePaymentElementsSavePaymentsEnabled').value === 'true') {
-        options.setup_future_usage = 'off_session';
-    }
-
-    window.stripePaymentElements = stripe.elements(options);
-
-    initStripePaymentElement();
+        initStripePaymentElement(response.customerEmail);
+    });
 }
 
 /* Stripe Payment Element */
