+++ cartridges/app_stripe_sfra/cartridge/templates/default/checkout/billing/paymentOptions.isml
@@ -16,19 +16,11 @@
      data-payment-method-id="CREDIT_CARD"
      data-is-new-payment="${pdict.customer.registeredUser && pdict.customer.customerPaymentInstruments.length ? false : true}"
 >
-	<isif condition="${!isCreditCardEnabled || !hasSavedCards}">
-		<input type="hidden" name="cardNumber" value="" class="cardNumber" id="cardNumber">
-		<div style="display:none;" id="new-card-form-container">
-		</div>
-	</isif>
-
-	<isloop items="${pdict.order.billing.payment.applicablePaymentMethods}" var="paymentOption">
-		<isif condition="${paymentOption.ID.equals('STRIPE_PAYMENT_REQUEST_BTN')}">
-			<div id="payment-request-button"><!-- A Stripe Element will be inserted here. --></div>
-			<isbreak/>
-		</isif>
-	</isloop>
-
+    <isif condition="${!isCreditCardEnabled || !hasSavedCards}">
+        <input type="hidden" name="cardNumber" value="" class="cardNumber" id="cardNumber">
+        <div style="display:none;" id="new-card-form-container">
+        </div>
+    </isif>
 
     <ul class="nav nav-tabs nav-fill payment-options" role="tablist">
         <isinclude template="checkout/billing/paymentOptions/paymentOptionsTabs" />
