+++ cartridges/app_stripe_sfra/cartridge/templates/default/checkout/expressCheckout/stripeMinicartExpressCheckoutButton.isml
@@ -1,0 +1,33 @@
+<isset name="stripeHelper" value="${require('*/cartridge/scripts/stripe/helpers/stripeHelper')}" scope="page" />
+<isset name="stripeEnabled" value="${stripeHelper.isStripeEnabled()}" scope="page" />
+<isset name="isStripeExpressCheckoutEnabledMinicart" value="${stripeHelper.isStripeExpressCheckoutMinicartEnabled()}" scope="page" />
+<isset name="stripePublicKey" value="${stripeHelper.getPublicApiKey()}" scope="page" />
+
+<isif condition="${stripeEnabled && isStripeExpressCheckoutMinicartEnabled && stripePublicKey}">
+    <isset name="stripeApiVersion" value="${stripeHelper.getStripeApiVersion()}" scope="page" />
+    <isset name="stripeOrder" value="${require('*/cartridge/scripts/stripe/helpers/checkoutHelper').getStripeOrderDetails(dw.order.BasketMgr.getCurrentBasket())}" scope="page" />
+    <isset name="stripeCheckoutHelper" value="${require('*/cartridge/scripts/stripe/helpers/checkoutHelper')}" scope="page" />
+    <isset name="stripeExpressCheckoutAppearance" value="${JSON.stringify(stripeHelper.getStripeExpressCheckoutAppearance())}" scope="page" />
+
+    <input type="hidden" id="stripePublicKey" value="${stripePublicKey}" />
+    <input type="hidden" id="stripeApiVersion" value="${stripeApiVersion}" />
+    <input type="hidden" id="stripe_express_checkout_url" value="${URLUtils.https("StripePayments-StripeQuickCheckout")}" />
+    <input type="hidden" id="stripe_return_url" value="${URLUtils.https("StripePayments-HandleAPM")}" />
+    <input type="hidden" id="stripe_get_shipping_options" value="${URLUtils.https("Stripe-GetShippingOptions")}" />
+    <input type="hidden" id="logStripeErrorMessageURL" value="${URLUtils.https('StripePayments-LogStripeErrorMessage')}" />
+    <input type="hidden" id="stripeCaptureMethod" value="${dw.system.Site.getCurrent().getCustomPreferenceValue('stripeChargeCapture') ? 'automatic' : 'manual'}" />
+    <input type="hidden" id="stripeExpressCheckoutAppearance" value="${stripeExpressCheckoutAppearance}" />
+    <input type="hidden" id="stripePaymentMethodsInBeta" value="${stripeHelper.getPaymentMethodsInBeta()}" />
+    <input type="hidden" id="stripe_order_amount" value="${stripeOrder ? stripeOrder.amount : 0}" />
+    <input type="hidden" id="stripe_order_currency" value="${stripeOrder && stripeOrder.currency ? stripeOrder.currency : session.currency.currencyCode}" />
+
+    <isscript>
+        var assets = require('*/cartridge/scripts/assets.js');
+        assets.addJs('/js/stripe.expressCheckout.js');
+    </isscript>
+
+    <isset name="displayExpressCheckout" value="true" scope="page" />
+
+    <div id="express-checkout-element-minicart" style="margin-top: 10px"></div>
+    <input type="hidden" name="${dw.web.CSRFProtection.getTokenName()}" value="${dw.web.CSRFProtection.generateToken()}" />
+</isif>
