+++ cartridges/int_stripe_sfra/cartridge/templates/default/checkout/billing/paymentOptions/stripePaymentElementContent.isml
@@ -7,5 +7,5 @@
             <div id="payment-element" style="width:100%" data-errormsg="${Resource.msg('apm.error.fail','stripe',null)}">
             </div>
         </div>
-  </fieldset>
+    </fieldset>
 </div>
